Update method of initializing actions
